+++ mslicer/src/task/mod.rs
@@ -16,6 +16,7 @@ mod project;
 mod remote_print;
 mod save_result;
 mod thread;
+mod webhook;
 pub use self::{
     acceleration_structures::BuildAccelerationStructures,
     file_dialog::FileDialog,
@@ -25,6 +26,7 @@ pub use self::{
     project::{ProjectLoad, ProjectSave},
     remote_print::{PrinterConnect, PrinterScan},
     save_result::SaveResult,
+    webhook::Webhook,
 };
 
 // Async operation that can be polled every frame.