+++ Changelog.md
@@ -24,6 +24,7 @@
 - Scale models before rotating
 - Encode file in async task
 - Fix crash when slicing without ever seeing workspace
+- Send custom webhooks on print completion
 
 ## v0.6.0 &mdash; March 23rd, 2026